requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9
Werkzeug>=3.0.1
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.2
gunicorn>=21.2.0
